requirements.txt
# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Vision Transformer Utilities
timm>=0.9.0
einops>=0.6.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Image Processing
pillow>=9.0.0

# Data Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Configuration and Utilities
pyyaml>=6.0
easydict>=1.10

# Progress Monitoring
tqdm>=4.65.0

# Model Export (Optional)
onnx>=1.12.0

# Dataset Handling
requests>=2.28.0

# Adversarial Robustness Testing (Optional)
pandas>=2.0.0  # For metrics logging

